#9758: Single pool submodule
